Advanced Platforming Part 1 - Fine tuning the mesh colliders for non-horizontal-surfaces (#62)
# Overview
Fine tunes the mesh colliders for non-horizontal-surfaces. As a step towards having sonic like behavior, this PR overhauls the previous rectangular collider shape with a composite of polygons for the penguin.

# Notes
Importantly, this does NOT yet address the issue of synchronizing the collider with the penguin shape, as for example, during walking and sliding animations the collider does NOT match the sprite at all, causing strange behaviors. That will be addressed in subsequent PRs.

# Changelog
* Apply brute-force rotation alignment (rigidbody up/forward axes as relative to surface normal)
* Replace flat ground segment with a spriteshape, using built in rock texture as a placeholder
* Sculpt ground shape to give a good variety of platform geometries for testing slope/rotational handling
* Tessellate all sprite components via adjusting physics outline in sprite editor
* Exclude texture/spriteshape related assets (currently trying out different ones, so there in gitignore for now)
* Replace box collider with a composite on the player object, with polygon colliders matching the physics outline for the head/torso/feet