Organizational: Unify verbage and ordering of fields in penguin controller (#73)
# Summary
* Vertically align assign statements in reset()
* Unify verbage and sort fields into a sensible ordering


# Context
Building off of the reorganization done in the last pr - [Inspector Attributes and Fields: Organization Improvements](https://github.com/jeffreypersons/Penguin-Quest/pull/72), this unifies the verbage (ie using thresholds in all cases instead of a random mix of tolerance/threshold/sensitivity/etc like before. Also, it reorders the constants to match the order they are used in fields (ie mass constants are at the end of the constant list, just like the attributes).

With dozens of constants, this makes things a lot easier to deal with and tweak via inspector, etc.