GroundChecker: Increase tolerated distance from ground to 5 units
This helps decrease the frequency of the issue where the penguin gets 'stuck' between two elevated points where the point below is out of reach.

In the future, there should be more than one sensor - at the front back, and middle, for both sliding and standing states respectively, with tests checking for differences between those. This would solve this 'stuck in a gap issue' as we would no longer have an over-reliance on the single ground check point directly beneath the penguin.